prov/efa: rename rdm_proto_v4.h to efa_rdm_protocol.h
This is to follow the file naming convention.
