+++ prov/efa/Makefile.include
@@ -106,7 +106,7 @@ _efa_headers = \
 	prov/efa/src/rdm/rxr_pkt_pool.h \
 	prov/efa/src/rdm/rxr_atomic.h \
 	prov/efa/src/rdm/efa_rdm_ope.h \
-	prov/efa/src/rdm/rdm_proto_v4.h \
+	prov/efa/src/rdm/efa_rdm_protocol.h \
 	prov/efa/src/rdm/rxr_tp_def.h \
 	prov/efa/src/rdm/rxr_tp.h \
 	prov/efa/src/rdm/efa_rdm_error.h \
